♻️ adjust entrypoint for creating atomFamilies -- fully cover silo
