+++ packages/atom.io/internal/src/families/index.ts
@@ -1,3 +1,4 @@
 export * from "./create-atom-family"
+export * from "./create-regular-atom-family"
 export * from "./create-readonly-selector-family"
 export * from "./create-selector-family"
