+++ packages/atom.io/src/silo.ts
@@ -2,9 +2,6 @@ import {
 	Store,
 	createAtom,
 	createAtomFamily,
-	createMutableAtom,
-	createMutableAtomFamily,
-	createRegularAtom,
 	createSelector,
 	createSelectorFamily,
 	createTimeline,
@@ -35,12 +32,7 @@ export class Silo {
 		const s = new Store(name, fromStore)
 		this.store = s
 		this.atom = (options) => createAtom(options, undefined, s)
-		this.atomFamily = (options) => {
-			if (`mutable` in options) {
-				return createMutableAtomFamily(options, s) as any
-			}
-			return createAtomFamily(options, s)
-		}
+		this.atomFamily = (options) => createAtomFamily(options, s)
 		this.selector = (options) => createSelector(options, undefined, s) as any
 		this.selectorFamily = (options) => createSelectorFamily(options, s) as any
 		this.transaction = (options) => createTransaction(options, s)
